Fix int to long for id to correspond to db big int

+++ src/main/java/com/example/cesizen/repository/CategorieRepository.java
@@ -3,5 +3,5 @@ package com.example.cesizen.repository;
 import com.example.cesizen.model.Categorie;
 import org.springframework.data.repository.CrudRepository;
 
-public interface  CategorieRepository extends CrudRepository<Categorie, Integer> {
+public interface  CategorieRepository extends CrudRepository<Categorie, Long> {
 }
